TPWDCEE-3688 change numeric conversation of basketMapper to real double type - the new PaymentSDK need this

+++ lib/WirecardEE/PaymentGateway/Mapper/BasketMapper.php
@@ -172,6 +172,6 @@ abstract class BasketMapper
      */
     public static function numberFormat($amount)
     {
-        return number_format($amount, 2, '.', '');
+        return (double)number_format($amount, 2, '.', '');
     }
 }
